# swagger.yml
swagger: '2.0'
x-authapikey: true
info:
  description: sample telco offer API
  version: '1.0'
  title: Telco Offer API
  license:
    name: For Code Challenge Use Only
basePath: /members/me
schemes:
  - https
paths:
  /login:
    post:
      tags:
        - auth
      summary: Generate a jwt token for a member
      description: Returns token to be used for other apis
      operationId: userLogin
      consumes:
        - application/json
      produces:
        - application/json
      parameters:
        - in: body
          name: body
          description: username and password of user
          required: true
          schema:
            type: object
            properties:
              username:
                type: string
              password:
                type: string
      responses:
        '200':
          description: successful operation
          schema:
            type: object
            properties:
              username:
                type: string
              token:
                type: string
        '500':
          description: Error in authentication
  /offers:
    get:
      tags:
        - offers
      summary: Get the offers for a member
      description: Returns list of offers available for a member
      operationId: getOffers
      produces:
        - application/json
      parameters:
        - in: header
          name: Authorization
          description: JWT Access Token
          type: string
          required: true
      responses:
        '200':
          description: successful operation
          schema:
            $ref: '#/definitions/OfferResponse'
        '400':
          description: Member not found to retrieve offers for
definitions:
  OfferResponse:
    type: object
    properties:
      CustomerOffers:
        type: array
        items:
          $ref: '#/definitions/CustomerOffer'
      Status:
        type: string
      ErrorInfo:
        $ref: '#/definitions/OfferErrorInfo'
  OfferErrorInfo:
    type: object
    properties:
      Code:
        type: string
      Source:
        type: string
      ErrorID:
        type: string
      Severity:
        type: string
      Message:
        type: string
  CustomerOffer:
    type: object
    properties:
      EndDate:
        type: string
      Hidden:
        type: string
      OfferId:
        type: string
      PropositionId:
        type: string
      StartDate:
        type: string
      Status:
        type: string
    title: CustomerOffer